Refactor component
