+++ src/feed/Feed.scss
@@ -1,0 +1,7 @@
+.Feed {
+  margin-top: 20px;
+}
+
+.Feed__content {
+  overflow-y: auto;
+}
