+++ src/store.js
@@ -18,7 +18,7 @@ import bookmarksReducer from './bookmarks/bookmarksReducer';
 import favoritesReducer from './favorites/favoritesReducers';
 import editorReducer from './post/Write/EditorReducers';
 import { responsiveReducer, mountResponsive } from './helpers/responsive';
-import reblogReducers from './app/reblog/reblogReducers';
+import reblogReducers from './app/Reblog/reblogReducers';
 import walletReducer from './wallet/walletReducer';
 
 export const messagesWorker = new MessagesWorker();
