--- src/styles/modules/post/post.scss
@@ -1,109 +1,0 @@
-.grid {
-  margin-top: 20px;
-
-  .grid-row-green {
-    .colored {
-      color: $green;
-    }
-  }
-
-  .grid-row-red {
-    .colored {
-      color: $red;
-    }
-  }
-
-  .grid-header {
-    a {
-      color: #7D7D8E;
-
-      &:hover {
-        color: $text-color;
-      }
-    }
-  }
-
-  .grid-content {
-    overflow-y: auto;
-  }
-
-  .grid-footer {
-    border-top: 1px solid rgba(0, 0, 0, 0.1);
-  }
-
-  .grid-row {
-    border: 1px solid #E7E7EC;
-    box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.05);
-    width: 100%;
-    border-radius: 5px;
-    max-width: 600px;
-    margin: 0 auto 20px;
-    table-layout: fixed;
-    background-color: #FFFFFF;
-  }
-
-  .cell {
-    vertical-align: middle;
-    overflow: hidden;
-
-    li i {
-      line-height: 47px;
-    }
-
-    li {
-      height: 47px;
-      line-height: 47px;
-      float: left;
-      vertical-align: middle;
-    }
-
-    li:last-child {
-      border-right: none;
-    }
-  }
-
-  .cell-top {
-    border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-
-    li {
-      padding: 0 1em;
-    }
-  }
-
-  .cell-body {
-    padding: 1em 1em;
-  }
-
-  .cell-bottom {
-    border-top: 1px solid rgba(0, 0, 0, 0.1);
-    text-align: center;
-
-    li {
-      width: 25%;
-      border-right: 1px solid rgba(0, 0, 0, 0.05);
-
-      &:hover {
-        background-color: #f7fbff;
-      }
-    }
-  }
-
-  h2 {
-    margin-bottom: 0.4em !important;
-  }
-}
-
-.thumbs {
-  background: white;
-  width: 100%;
-  text-align: center;
-  overflow: hidden;
-  line-height: 0;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-}
-
-@media (max-width: 600px) {
-  .grid-row {
-    border-width: 1px 0 1px 0 !important;
-  }
-}
