--- src/styles/modules/post/replies.scss
@@ -1,7 +1,0 @@
-.replies {
-  li {
-    border-top: 1px solid rgba(0, 0, 0, 0.1);
-    padding: 1em;
-    word-break: break-word;
-  }
-}
